scripts/requirements.txt
numpy==1.19.4
pandas==1.1.5
pysam==0.16.0.1
python-dateutil==2.8.1
pytz==2020.4
PyVCF==0.6.8
six==1.15.0
